move tracksWithChaptersOrVideosSlugs near the validator + notes on restrictions
